Added gRPC + protobuf. Need to fix the ProjectService/(ProjectService + AuthService)'s compose file

+++ backend/project-service/src/main/java/ru/itpark/projectservice/infrastructure/repositories/ProjectRepo.java
@@ -5,6 +5,6 @@ import org.springframework.stereotype.Repository;
 import ru.itpark.projectservice.domain.Project;
 
 @Repository
-public interface ProjectRepo extends JpaRepository<Project, Integer> {
+public interface ProjectRepo extends JpaRepository<Project, Long> {
 
 }
